+++ examples/complete/data.tf
@@ -1,5 +1,20 @@
+data "local_file" "json_file" {
+  filename = "${path.module}/files/policy.json"
+}
 
 data "aws_iam_policy_document" "stackset_administration_role_assume_role_policy" {
+  statement {
+    actions = ["sts:AssumeRole"]
+    effect  = "Allow"
+
+    principals {
+      identifiers = ["cloudformation.amazonaws.com"]
+      type        = "Service"
+    }
+  }
+}
+
+data "aws_iam_policy_document" "stack_assume_role_policy" {
   statement {
     actions = ["sts:AssumeRole"]
     effect  = "Allow"
@@ -15,7 +30,7 @@ data "aws_iam_policy_document" "stackset_administration_role_execution_policy" {
   statement {
     actions   = ["sts:AssumeRole"]
     effect    = "Allow"
-    resources = ["arn:aws:iam::*:role/${local.stackset_name}-execution-role"]
+    resources = ["arn:aws:iam::*:role/${var.name}set-execution-role"]
   }
 }
 
@@ -30,6 +45,7 @@ data "aws_iam_policy_document" "stackset_execution_role_assume_role_policy" {
     }
   }
 }
+
 
 data "aws_iam_policy_document" "stackset_execution_role_example_policy" {
   #checkov:skip=CKV_AWS_111 "Ensure IAM policies does not allow write access without constraints"