+++ examples/complete/locals.tf
@@ -1,4 +1,1 @@
-locals {
-  stackset_name = "complete-stackset-example"
-  tags          = merge({ Name = local.stackset_name }, var.tags)
-}
+# Empty